agent/requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
strands-agents
strands-agents-tools
langfuse
ragas
langchain-aws
pandas
datasets
flask